msm_serial_hs: Fix race between mod_timer and del_timer calls
Deletion of timer and synchronization of the handler will happen
during del_timer, but in the issue case the timer got deleted and
before handler_sync should happen the timer is getting modified or
accessed which further leads to kernel panic.

Fix this by modifying del_timer to del_timer_sync so that the timer
is protected until synchronization happen.

Change-Id: If5dec66b9bbcb5369898536ba1c727b487a13018
Signed-off-by: Venkata Manasa Kakarla <venkka@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>