Revert "msm: kgsl: Mark IRQ and worker thread as performance critical"
This reverts commit 1ae0f60.

This is wrong. This causes 3 IRQs to be affined to the big CPU cluster,
not just the primary kgsl_3d0_irq one. As a result, the perf crit API
thinks that the 2 extra IRQs are critical and will balance them despite
them being rarely used (kgsl_hfi_irq and kgsl_gmu_irq).

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>